+++ test/modules/Auction/MockAuctionModule.sol
@@ -30,7 +30,7 @@ contract MockAuctionModule is AuctionModule {
         uint96 id_,
         uint96 amount_,
         bytes calldata auctionData_
-    ) internal override returns (uint256 payout, bytes memory auctionOutput) {}
+    ) internal override returns (uint96 payout, bytes memory auctionOutput) {}
 
     function _bid(
         uint96 id_,
@@ -42,7 +42,7 @@ contract MockAuctionModule is AuctionModule {
 
     function _settle(uint96 lotId_) internal override returns (Settlement memory, bytes memory) {}
 
-    function _claimProceeds(uint96 lotId_) internal override returns (uint256, uint256, uint256) {}
+    function _claimProceeds(uint96 lotId_) internal override returns (uint96, uint96, uint96) {}
 
     function _refundBid(
         uint96 lotId_,
