+++ test/modules/Auction/MockBatchAuctionModule.sol
@@ -71,7 +71,7 @@ contract MockBatchAuctionModule is AuctionModule {
         uint96,
         uint96,
         bytes calldata
-    ) internal pure override returns (uint256, bytes memory) {
+    ) internal pure override returns (uint96, bytes memory) {
         revert Auction_NotImplemented();
     }
 
@@ -144,7 +144,7 @@ contract MockBatchAuctionModule is AuctionModule {
         return (lotSettlements[lotId_], "");
     }
 
-    function _claimProceeds(uint96 lotId_) internal override returns (uint256, uint256, uint256) {
+    function _claimProceeds(uint96 lotId_) internal override returns (uint96, uint96, uint96) {
         // Update status
         lotStatus[lotId_] = Auction.Status.Claimed;
